refactor: move eacces to its own module
